Implemented a tab switch for different views of the metadata - the original hierarchical tabbed view, and the newer grid-based page view.
The latter implementation is currently partial. There are questions to be answered on
how we wish to handle linking (clicks - jsGrid supports callbacks to cell clicks) as
well as metadata hierarchies and presentation. As such only basic fields are presented
in the grid-based view until such time these design issues are resolved.